+++ src/hooks/review/useAddReview.js
@@ -1,17 +1,8 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
-import {
-  updateDoc,
-  doc,
-  increment,
-  addDoc,
-  collection,
-  serverTimestamp,
-} from "firebase/firestore";
-import { db } from "@/services/firebase";
-import { uploadImage } from "@/utils/uploadImage";
 import { QUERY_KEYS } from "@/constants/queryKeys";
+import { createReview } from "@/services/review/createReview";
 
 /**
  * 리뷰 추가 기능을 제공하는 커스텀 훅
@@ -21,8 +12,8 @@ import { QUERY_KEYS } from "@/constants/queryKeys";
  * @returns {Object} 리뷰 추가 관련 함수와 상태
  */
 export default function useAddReview({
-  onSuccessCallback = () => {},
-  onErrorCallback = () => {},
+  onSuccessCallback,
+  onErrorCallback,
 } = {}) {
   const navigate = useNavigate();
   const user = useSelector((state) => state.user.user);
@@ -43,92 +34,18 @@ export default function useAddReview({
   } = useMutation({
     mutationFn: async (review) => {
       checkAuth();
-
-      const {
-        title,
-        destination,
-        content,
-        rating,
-        image,
-        type = "standard",
-        interviewAnswers,
-        interviewQuestions,
-      } = review;
-
-      if (!title?.trim()) throw new Error("제목을 입력해주세요.");
-      if (!destination?.trim()) throw new Error("여행지명을 입력해주세요.");
-      if (!rating) throw new Error("별점을 입력해주세요.");
-      if (type === "standard" && (!content?.trim() || content.length < 100)) {
-        throw new Error("내용을 100자 이상 입력해주세요.");
-      }
-
-      let imageUrl = "";
-      if (image instanceof File) {
-        const supportedTypes = ["image/jpeg", "image/png", "image/gif"];
-        if (!supportedTypes.includes(image.type)) {
-          throw new Error("지원되는 이미지 형식은 JPEG, PNG, GIF입니다.");
-        }
-
-        try {
-          imageUrl = await uploadImage(image, "reviews", user.uid);
-        } catch (err) {
-          throw new Error(err.message || "이미지 업로드에 실패했습니다.");
-        }
-      } else if (typeof image === "string") {
-        imageUrl = image; // 수정 시 기존 이미지 URL 유지
-      } else if (image) {
-        throw new Error("올바르지 않은 이미지 형식입니다.");
-      }
-
-      const reviewData = {
-        title,
-        destination,
-        rating,
-        imageUrl,
-        createdBy: {
-          uid: user.uid,
-          displayName: user.displayName || "익명",
-          photoURL: user.photoURL || "",
-        },
-        createdAt: serverTimestamp(),
-        likeCount: 0,
-        commentCount: 0,
-        type,
-      };
-
-      if (type === "standard") {
-        reviewData.content = content;
-      } else {
-        reviewData.interviewAnswers = interviewAnswers;
-        reviewData.interviewQuestions = interviewQuestions;
-      }
-
-      const docRef = await addDoc(collection(db, "reviews"), reviewData);
-
-      await updateDoc(doc(db, "users_private", user.uid), {
-        reviewCount: increment(1),
-      });
-
-      return docRef.id;
+      return await createReview(review, user);
     },
-
     onSuccess: (newId) => {
       alert("리뷰가 성공적으로 등록되었습니다!");
       queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.REVIEWS] });
-      onSuccessCallback(newId);
+      onSuccessCallback?.(newId);
     },
-
     onError: (error) => {
-      console.error(error);
       alert(`리뷰 등록 중 오류가 발생했습니다: ${error.message}`);
-      onErrorCallback(error);
+      onErrorCallback?.(error);
     },
   });
 
-  return {
-    addReview,
-    isLoading,
-    isError,
-    error,
-  };
+  return { addReview, isLoading, isError, error };
 }
